@@ -1,4 +1,6 @@
 #!/usr/bin/env python3
+"""Exporter classes for PeakRDL-sv."""
+
 from __future__ import annotations
 
 import argparse
@@ -6,17 +8,25 @@ from pathlib import Path
 
 from mako.template import Template
 from pkg_resources import resource_filename
-from systemrdl.node import AddrmapNode
-from systemrdl.node import RootNode
+from systemrdl.node import AddrmapNode, RootNode
 
 from peakrdl_sv.listener import Listener
 
 
 class VerilogExporterBase:
-    def __init__(self):
+    """A SystemRDL to Verilog exporter."""
+
+    def __init__(self) -> None:
+        """Initialise the exporter with a listener."""
         self.listener = Listener()
 
     def walk(self, node: RootNode | AddrmapNode) -> None:
+        """Walk an addr map from a root node.
+
+        Args:
+          node: RootNode | AddrmapNode: The root node of the map, or the map itself
+
+        """
         if isinstance(node, RootNode):
             node = node.top
 
@@ -27,6 +37,13 @@ class VerilogExporterBase:
         node: RootNode | AddrmapNode | None,
         options: argparse.Namespace,
     ) -> None:
+        """Export a register map to (System)Verilog.
+
+        Args:
+          node: RootNode | AddrmapNode | None: The root node of the PeakRDL reg map
+          options: argparse.Namespace: the output path
+
+        """
         if node:
             self.walk(node)
         elif not self.top:
@@ -53,10 +70,19 @@ class VerilogExporterBase:
 
 
 class PythonExporterBase:
-    def __init__(self):
+    """A simple exporter for RDL -> Python."""
+
+    def __init__(self) -> None:
+        """Initialise the exporter with a listener."""
         self.listener = Listener()
 
     def walk(self, node: RootNode | AddrmapNode) -> None:
+        """Walk an addr map from a root node.
+
+        Args:
+          node: RootNode | AddrmapNode: The root node of the map, or the map itself
+
+        """
         if isinstance(node, RootNode):
             node = node.top
 
@@ -67,7 +93,14 @@ class PythonExporterBase:
         node: RootNode | AddrmapNode | None,
         options: argparse.Namespace,
     ) -> None:
-        if node:
+        """Export a register map to Python.
+
+        Args:
+          node: RootNode | AddrmapNode | None: The root node of the PeakRDL reg map
+          options: argparse.Namespace: the output path
+
+        """
+        if node is not None:
             self.walk(node)
         elif not self.top:
             raise RuntimeError
