@@ -1,40 +1,52 @@
+"""Provide classes for modelling register access in a coco.tb simulation."""
+
 from __future__ import annotations
 
 import logging
+from collections.abc import Generator
 from dataclasses import dataclass
 from typing import Any
-from typing import Generator
 
 import numpy as np
 from systemrdl import RDLCompiler
 
 from peakrdl_sv.callbacks import CallbackSet
 from peakrdl_sv.listener import Listener
-from peakrdl_sv.node import AddressMap
-from peakrdl_sv.node import Field
-from peakrdl_sv.node import Register
+from peakrdl_sv.node import AddressMap, Field, Register
 
 
 @dataclass
 class FieldWrapper:
-    """Wraps a Field by storing along side it a (desired) value"""
+    """Wraps a Field by storing along side it a (desired) value."""
 
     field: Field
     value: int
 
 
 class RegModel:
-    """Register Abstraction Layer Model
+    """Register Abstraction Layer Model.
 
-    :param rdlfile: The rdl file from which we create the model
-    :type rdlfile: str
-    :param callbacks: The set of coco.tb callbacks to perform read/write operations
-    :type callbacks: CallbackSet
-    :param log: A logger
-    :type log: Any
+    This class provides an interface for interacting with a SystenRFL register map
+    within an RTL simulation
+
     """
 
-    def __init__(self, rdlfile: str, callbacks: CallbackSet, log, debug: bool = False):
+    def __init__(
+        self,
+        rdlfile: str,
+        callbacks: CallbackSet,
+        log: logging.Logger,
+        debug: bool = False,
+    ) -> None:
+        """Initialise the RAL model.
+
+        Args:
+            rdlfile(str): The rdl file from which we create the model
+            callbacks(CallbackSet): The set of coco.tb callbacks to perform read/write operations
+            log(logging.Logger): A logger
+            debug (bool): If true, set log level to debug.
+
+        """  # noqa: E501
         self._callbacks = callbacks
         self._top_node: AddressMap = self._parse_rdl(rdlfile)
         self._reg_map = self._map_registers()
@@ -49,6 +61,15 @@ class RegModel:
     # --------------------------------------------------------------------------------
 
     def _parse_rdl(self, rdlfile: str) -> AddressMap:
+        """Parse the RDL file into an AddressMap object.
+
+        Args:
+            rdlfile (str): Path to the address map
+
+        Returns:
+            AddressMap: The generated PeakRDL-sv AddressMap
+
+        """
         rdlc = RDLCompiler()
         rdlc.compile_file(rdlfile)
         root = rdlc.elaborate()
@@ -57,23 +78,26 @@ class RegModel:
         return listener.top_node
 
     def _map_registers(self) -> dict[str, Register]:
-        """Builds a dictionary which maps register relative paths to the register
-        itself, from which its properties may be extracted
+        """Build a dictionary which maps register relative paths to the register.
 
-        :return: A map of register relative paths to Register objects
-        :rtype: Dict[str, Register]
+        Returns:
+          Dict[str, Register]: A map of register relative paths to Register objects
+
         """
         return {
             register.path("."): register for register in self._top_node.get_registers()
         }
 
     def _map_fields(self) -> dict[str, dict[str, FieldWrapper]]:
-        """Maps register names to dictionaries of fields which maps to (desired) values
+        """Map register names to dictionaries of fields which maps to (desired) values.
 
-        :raises AttributeError: _map_fields called before _map_registers
-        :return: A dict of {register names : dict of {field names : (field, value)}}
-        :rtype: Dict[str, Dict[str, FieldWrapper]]
-        """
+        Returns:
+          Dict[str, Dict[str, FieldWrapper]]: A dict of {register names : dict of {field names : (field, value)}}
+
+        Raises:
+          AttributeError: _map_fields called before _map_registers
+
+        """  # noqa: E501
         try:
             return {
                 reg_name: {
@@ -92,35 +116,39 @@ class RegModel:
     # --------------------------------------------------------------------------------
 
     def get_register_by_name(self, reg_name: str) -> Register | None:
-        """Gets a register by its relative path name, '.' separated per tier of the
-        hierarchy
+        """Get a register by its relative path name.
 
-        :param reg_name: The register name as a . separated string
-        :type reg_name: str
-        :return: The target register, None if the register name is not found in the
-        regmap
-        :rtype: Register | None
+        The reg_name must be '.' separated per tier of the hierarchy.
+
+        Args:
+          reg_name(str): The register name as a . separated string
+
+        Returns:
+          Register | None: The target register, None if the register name is not found
+
         """
         try:
             return self._reg_map[reg_name]
         except KeyError as e:
             raise KeyError(f"Could not find register in {self._reg_map.keys()}") from e
 
-    def get_field_by_name(self, reg_name: str, field_name: str) -> FieldWrapper | None:
-        """Gets a field by its relative path
+    def get_field_by_name(self, reg_name: str, field_name: str) -> FieldWrapper:
+        """Get a field by its relative path.
 
-        :param reg_name: The register name as a . separated string
-        :type reg_name: str
-        :param field_name: The field name within the register
-        :type field_name: str
-        :return: The target field and its desired value as a FieldWrapper object
-        :rtype: FieldWrapper | None
+        Args:
+          reg_name(str): The register name as a . separated string
+          field_name(str): The field name within the register
+
+        Returns:
+          FieldWrapper: The target field and its desired value as a FieldWrapper
+
         """
         try:
             return self._desired_values[reg_name][field_name]
         except KeyError as e:
             all_keys = (
-                self._desired_values.keys() + self._desired_values[reg_name].keys()
+                *self._desired_values.keys(),
+                *self._desired_values[reg_name].keys(),
             )
             raise KeyError(f"Could not find field in {all_keys}") from e
 
@@ -129,17 +157,16 @@ class RegModel:
         target: Register,
         value: int,
     ) -> Generator[tuple[Any, int], Any, None]:
-        """Split a value over the fields of a register with correct masking depending
-        on the field location
+        """Split a value over the fields of a register with correct masking depending on the field location.
 
-        :param target: The register to target
-        :type target: Register
-        :param value: The value to write to the registers
-        :type value: int
-        :yield: tuples of the field address and the value to write
-        :rtype: Generator[tuple[Any, int], Any, None]
-        """
+        Args:
+          target(Register): The register to target
+          value(int):  The value to write to the registers
 
+        Yields:
+            tuples of the field address and the value to write)
+
+        """  # noqa: E501
         # format the write value as a `regwidth`-bit binary string
         bits = f"{value:0{target.regwidth}b}"
 
@@ -150,7 +177,7 @@ class RegModel:
             upper = target.regwidth - field.lsb
 
             # perform the slice (python slicing non inclusive)
-            yield (field.absolute_address, int(bits[lower:upper], 2))  # noqa: E203
+            yield (field.absolute_address, int(bits[lower:upper], 2))
 
     # --------------------------------------------------------------------------------
     # Generic read and write wrappers
@@ -160,19 +187,20 @@ class RegModel:
         self,
         reg_name: str,
         data: int | None = None,
-        **kwargs,
+        **kwargs: Any,  # noqa: ANN401
     ) -> None:
-        """Writes the value of the DUT register. Two options for calling:
+        """Write the value of the DUT register.
 
+        Two options for calling:
         (1) register name + data -> write to named register a particular value
         (2) register name -> write desired value to a named register
 
-        :param reg_name: Register name to write to
-        :type reg_name: str | int
-        :param data: Data to write to register, defaults to None
-        :type data: int | None, optional
-        """
+        Args:
+            reg_name (str | int): Register name to write to
+            data (int | None): Data to write to register, defaults to None
+            kwargs (Any): kwargs to pass to async_write_callback
 
+        """
         self._log.debug(f"Performing write to target register {reg_name}")
 
         target = self.get_register_by_name(reg_name)
@@ -205,16 +233,15 @@ class RegModel:
             self.set(reg_name, data)
 
     async def read(self, reg_name: str, field_name: str | None = None) -> int:
-        """Reads the value of a register or field from the DUT
+        """Read the value of a register or field from the DUT.
 
-        :param reg_name: The register name to read from
-        :type reg_name: str
-        :param field_name: The field name to read from, defaults to None
-        :type field_name: str | None, optional
-        :return: The value read from the DUT
-        :rtype: int
+        reg_name (str): The register name to read from
+        field_name (str | None): The field name to read from, defaults to None
+
+        Returns:
+            int: The value read from the DUT
+
         """
-
         self._log.debug(f"Performing read from target register {reg_name}.{field_name}")
 
         target = self.get_register_by_name(reg_name)
@@ -224,8 +251,7 @@ class RegModel:
         )
 
         async def read_register(target: Register) -> int:
-            """Reads a registers in `accesswidth` sized chunks"""
-
+            """Readsa registers in `accesswidth` sized chunks."""
             result = 0
             # if its wide we need to read in accesswidth chunks
             for subreg in range(target.subregs):
@@ -236,15 +262,12 @@ class RegModel:
             return result
 
         async def read_field(target: Register, field_name: str) -> int:
-            """Reads an individual field from a regster and references it to 0
+            """Read an individual field from a regster and references it to 0.
 
             Assumes field width is limited by access width
             """
-
             # find the field
             target_field = self.get_field_by_name(target.name, field_name).field
-            if target_field is None:
-                raise ValueError(f"Could not find field in register ({target.name})")
 
             # perform the read
             addr = target_field.absolute_address
@@ -265,7 +288,7 @@ class RegModel:
     # --------------------------------------------------------------------------------
 
     def set(self, reg_name: str, value: int) -> None:
-        """Sets a the desired value of a register
+        """Set a the desired value of a register.
 
         This method will split the passed in value over the bits of the field
         i.e. if you pass in 0xE5 == 0b11101010 and the register has the format
@@ -275,10 +298,10 @@ class RegModel:
         }
         then we'd have data0 = 2b10, data1 = 4b1110
 
-        :param reg_name: The target register's name
-        :type reg_name: str
-        :param value: The value to write to the register
-        :type value: int
+        Args:
+          reg_name(str): The target register's name
+          value(int): The value to write to the register
+
         """
         self._log.debug(f"Setting reg ({reg_name}) with value {value}")
 
@@ -294,14 +317,13 @@ class RegModel:
             self._desired_values[reg_name][field.inst_name].value = masked_values[idx]
 
     def set_field(self, reg_name: str, field_name: str, value: int) -> None:
-        """Sets the desired value of a field
+        """Set the desired value of a field.
 
-        :param reg_name: The target register's name
-        :type reg_name: str
-        :param field_name: The name of the field to update
-        :type field_name: str
-        :param value: the value to write to the field
-        :type value: int
+        Args:
+          reg_name(str): The target register's name
+          field_name(str): The name of the field to update
+          value(int): the value to write to the field
+
         """
         self._log.debug(f"Setting field ({reg_name}.{field_name}) with value {value}")
 
@@ -314,12 +336,15 @@ class RegModel:
             ) from e
 
     def get(self, reg_name: str) -> int:
-        """Gets the desired value of a register
+        """Get the desired value of a register.
 
-        :param reg_name: The target register's name
-        :type reg_name: str
-        :return: the register's value
-        :rtype: int
+        Args:
+          reg_name(str): The target register's name
+          reg_name: str:
+
+        Returns:
+          int: the register's value
+
         """
         self._log.debug(f"Getting reg ({reg_name}) value")
 
@@ -334,14 +359,17 @@ class RegModel:
         return value
 
     def get_field(self, reg_name: str, field_name: str) -> int:
-        """Gets the desired value of a field
+        """Get the desired value of a field.
 
-        :param reg_name: The target register's name
-        :type reg_name: str
-        :param field_name: The name of the field to update
-        :type field_name: str
-        :return: The field's value
-        :rtype: int
+        Args:
+          reg_name(str): The target register's name
+          field_name(str): The name of the field to update
+          reg_name: str:
+          field_name: str:
+
+        Returns:
+          int: The field's value
+
         """
         self._log.debug(f"Getting field ({reg_name}.{field_name}) value")
 
@@ -353,23 +381,27 @@ class RegModel:
                 f" ({self._desired_values.keys()})",
             ) from e
 
-    def randomize(self, reg_name: str):
-        """Randomizes the value of a register
+    def randomize(self, reg_name: str) -> None:
+        """Randomizes the value of a register.
 
-        :param reg_name: The name of the register to randomise
-        :type reg_name: str
+        Args:
+          reg_name(str): The name of the register to randomise
+          reg_name: str:
+
         """
         target = self.get_register_by_name(reg_name)
 
         self.set(reg_name, np.random.randint(0, 1 << target.regwidth))
 
-    def randomize_field(self, reg_name: str, field_name: str):
-        """Randomizes the value of a field in a register
+    def randomize_field(self, reg_name: str, field_name: str) -> None:
+        """Randomizes the value of a field in a register.
 
-        :param reg_name: The name of the parent register
-        :type reg_name: str
-        :param field_name: The name of the field to randomize
-        :type field_name: str
+        Args:
+          reg_name(str): The name of the parent register
+          field_name(str): The name of the field to randomize
+          reg_name: str:
+          field_name: str:
+
         """
         target_field = self.get_field_by_name(reg_name, field_name)
 
@@ -380,10 +412,12 @@ class RegModel:
         )
 
     def reset(self, reg_name: str) -> None:
-        """Resets the desired value of a register
+        """Reset the desired value of a register.
 
-        :param reg_name: The register you wish to reset
-        :type reg_name: str
+        Args:
+          reg_name(str): The register you wish to reset
+          reg_name: str:
+
         """
         target = self.get_register_by_name(reg_name)
 
@@ -392,12 +426,14 @@ class RegModel:
             self._desired_values[reg_name][field.inst_name].value = field.reset or 0
 
     def reset_field(self, reg_name: str, field_name: str) -> None:
-        """Resets the desired value of a field within a register
+        """Reset the desired value of a field within a register.
 
-        :param reg_name: The register you wish to target
-        :type reg_name: str
-        :param field_name: The field you wish to reset
-        :type field_name: str
+        Args:
+          reg_name(str): The register you wish to target
+          field_name(str): The field you wish to reset
+          reg_name: str:
+          field_name: str:
+
         """
         target = self._desired_values[reg_name][field_name]
 
